% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_filter.R
\name{get_edge_disparity_pvals}
\alias{get_edge_disparity_pvals}
\title{Disparity filter for an undirected weighted network}
\usage{
get_edge_disparity_pvals(net, deg.one.pval = 1)
}
\arguments{
\item{net}{igraph; An undirected weighted network.}

\item{deg.one.pval}{numeric; By default, edges of degree-one nodes are assigned a p-value of 1. This can be changed through this parameter.}
}
\value{
A numeric vector of length \code{ecount(net)} with a p-value for each weighted edge of the input network.
}
\description{
For each weighted edge in the given network, it computes two p-values representing its local significance for the two nodes it touches.
This function returns only the minium p-value for the two that are computed.
}
\examples{
# Get disparity p-values for the edges of the included US Airports network
edge.pvals <- get_edge_disparity_pvals(net = air)

}
\references{
Serrano, M. A. et al. (2009) Extracting the multiscale backbone of complex weighted networks. \emph{PNAS} 106(16).
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
