\name{network}
\docType{data}
\alias{network}
\title{A Small Sample Network}
\description{
This data set gives the edgelist of a small sample network.
}
\usage{network}
\format{\code{network} is data frame containing 300 observations (rows) representing the edges of a small sample network, and 3 variables (columns) named \code{from}, \code{to}, and \code{weight} representing the vertices and the weight of the edges.}
\keyword{datasets}
