% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_filter.R
\name{plot_degree_vs_disparity}
\alias{plot_degree_vs_disparity}
\title{Visualise node disparities as a function of node degree}
\usage{
plot_degree_vs_disparity(net, node.disp)
}
\arguments{
\item{net}{igraph; An undirected weighted network.}

\item{node.disp}{numeric; Numeric vector of length \code{vcount(net)} with the disparity measure of each node in the network.}
}
\value{
A ggplot comparing node disparities as a function of node degree.
}
\description{
Given an undirected weighted network and node disparities precomputed with function \code{get_node_disparity}, 
generates a ggplot comparing node disparities as a function of node degree.
Reference lines for null disparities, perfect node homogeneity and perfect node heterogeneity are also shown.
}
\examples{
# Compute the disparity of the nodes in the included US Airports network
disp <- get_node_disparity(net = air)
# Plot degree vs disparities
p <- plot_degree_vs_disparity(net = air, node.disp = disp)

}
\references{
Serrano, M. A. et al. (2009) Extracting the multiscale backbone of complex weighted networks. \emph{PNAS} 106(16).
}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
